docker/worker/requirements.txt
llama-cpp-python
redis
    